Nasa-hackthon-final/farm_navigator_backend/requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
httpx
python-dotenv==1.0.0